requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-google-genai
PyPDF2
pinecone-client  # <<< MUDE AQUI: remova o "==2.2.4"
python-dotenv
python-multipart